app/requirements.txt
# Factory GitHub App dependencies
# Per briefing/apps/authentication.md

fastapi==0.109.0
uvicorn[standard]==0.25.0
httpx==0.26.0
pyjwt[crypto]==2.8.0
python-dotenv==1.0.0
githubkit==0.11.0